Refactored code for s3 enumeration
