+++ cloudhunter/internal/shared/helpers.go
@@ -2,6 +2,7 @@ package shared
 
 import (
 	"encoding/json"
+	"fmt"
 	"log"
 	"net/url"
 )
@@ -24,4 +25,15 @@ func ParseJsonPolicyDocument(policyData string) string {
 	}
 
 	return string(policy)
+}
+
+func RenderDirectoryTree(nodes []*S3Node, indent string) {
+	for _, node := range nodes {
+		if node.IsFolder {
+			fmt.Printf("%s %s\n", indent, node.Name)
+			RenderDirectoryTree(node.Children, indent+"  ")
+		} else {
+			fmt.Printf("%s %s\n", indent, node.Name)
+		}
+	}
 }
