+++ cloudhunter/internal/shared/types.go
@@ -1,0 +1,7 @@
+package shared
+
+type S3Node struct {
+	Name     string
+	IsFolder bool
+	Children []*S3Node
+}
